Refactor navigation links and project section
Update navigation links in MainComponent.vue to use "Mes réalisations" instead of "Mes projets". Also, update the id of the project section to "realisations" in both MainComponent.vue and ProjetComponent.vue.

Update project description in ProjetComponent.vue to use "réalisations" instead of "projets". Also, correct the verb form in the description to "Contribuez" instead of "Contribué".

These changes improve the clarity and consistency of the navigation and project section labels.